more file renaming and restructuring
