+++ rentme/src/App.js
@@ -5,7 +5,7 @@ import Login from "./components/Login";
 import ManagerLogin from "./components/ManagerLogin";
 import { ManagerCard, ManagerDash } from "./components/propertyOwner/manager";
 
-import { PropertyDash, PropertyCard, addManagerForm, VendorAddressBk, TenantAddressBk, addTenantForm } from "./components/propertyOwner";
+import { PropertyDash, PropertyCard, addManagerForm, VendorAddressBK, TenantAddressBk, addTenantForm } from "./components/propertyOwner";
 import { RentReceipts, IssueForm} from "./components/tenant";
 import PropertyOwnerSignup from "./components/PropertyOwnerSignup";
 import RenterSignUp from "./components/RenterSignUp";
@@ -25,27 +25,33 @@ const App = () => (
 
         <hr />
 
+        {/* Authentication/Account Related */}
         <Route exact path={ROUTES.LOGIN} component={Login} />
         <Route exact path={ROUTES.RENTER_SIGNUP} component={RenterSignUp} />
         <Route exact path={ROUTES.OWNER_SIGNUP} component={PropertyOwnerSignup} />
         <Route exact path={ROUTES.PASSWORD_FORGET} component={PasswordForget} />
         <Route exact path={ROUTES.ACCOUNT} component={Account} />
 
+        {/* Dashboard Views */}
         <Route exact path={ROUTES.TENANT_DASHBOARD} component={TenantDashboard} />
         <Route exact path={ROUTES.OWNER_DASHBOARD} component={PropertyDash} />
 
+        {/* Property Owner Routes */}
         <Route exact path={ROUTES.PROPERTY_CARD} component={PropertyCard} />
+        <Route exact path={ROUTES.ADD_TENANT} component={addTenantForm} />
+        <Route exact path={ROUTES.TENANT_ADDRESSBK} component={TenantAddressBk} />
+        <Route exact path={ROUTES.VENDOR_ADDRESSBK} component={VendorAddressBK} />
 
+        {/* Tenant Routes */}
+        <Route exact path={ROUTES.RENT_RECEIPTS} component={RentReceipts} />
+        <Route exact path={ROUTES.ISSUE_FORM} component={IssueForm} />
+
+
+        {/* Hold for Stretch */}
         <Route path="/manager-login" component={ManagerLogin} />
         <Route path="/manager-dash" component={ManagerDash} />
-        <Route path="/vendor-addbook" component={VendorAddressBk} />
-        <Route path="/tenant-addbook" component={TenantAddressBk} />
         <Route path="/add-manager" component={addManagerForm} />
         <Route path="/manager-profile/:id" component={ManagerCard} />
-        <Route path="/view-receipts" component={RentReceipts} />
-        <Route path="/issue-report" component={IssueForm} />
-        <Route path="/property-card/:id" component={PropertyCard} />
-        <Route path="/add-tenant" component={addTenantForm} />
       </>
 
     </Router>
