+++ rentme/src/components/propertyOwner/index.js
@@ -3,7 +3,7 @@ import PropertyCard from "./PropertyCard";
 import addManagerForm from "./addManagerForm";
 import VendorAddressBk from "./VendorAddressBk";
 import TenantAddressBk from "./TenantAddressBk";
-import addTenantForm from "./addTenantForm";
+import AddTenantForm from "./AddTenantForm";
 
 
-export { PropertyDash, PropertyCard, addManagerForm, VendorAddressBk, TenantAddressBk, addTenantForm }
+export { PropertyDash, PropertyCard, addManagerForm, VendorAddressBk, TenantAddressBk, AddTenantForm }
