+++ rentme/src/constants/routes.js
@@ -9,3 +9,9 @@ export const TENANT_DASHBOARD = '/tenant-dash';
 export const OWNER_DASHBOARD = '/owner-dash';
 
 export const PROPERTY_CARD = '/property-card/:id';
+export const ADD_TENANT = '/add-tenant';
+export const TENANT_ADDRESSBK = '/tenant-addbook';
+export const VENDOR_ADDRESSBK = '/vendor-addbook';
+
+export const RENT_RECEIPTS = '/view-receipts';
+export const ISSUE_FORM = '/issue-report';
